Eliminate useless EditorialContentApi interface
